requirements.txt
python-telegram-bot==20.0
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0
